:art:Refactor: refactor atoms file atoms 파일 위치 변경
