+++ coin/src/components/units/main/chart/CoinChart.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useRecoilState } from 'recoil';
-import { darkMode } from '../../../commons/atoms';
+import { darkMode } from '../../../commons/util/atoms';
 
 declare global {
     interface Window {
